Factor out tree logic for easier refactoring
